Switch to C++11.  Use enum class for Type enum.

+++ comm.h
@@ -53,7 +53,7 @@ struct ServerMsg {
 /* Sockets for communication. */
 extern int server_socket;
 /* Current property. */
-extern size_t current_property;
+extern int current_property;
 
 
 #endif /* COMM_H */
